Substitutes AtomicIntegers for the pointers.
This is probably overkill, but these shared pointers should now be
thread safe. The real advance here is using the modulus operator: it
doesn't actually matter what the value of the counter is, or whether
you double-increment in some cases, it only matters that you don't
index beyond the array boundary. So we could probably revert this to
non-volatile ints, and just increment them forever. Turns out, though,
that AtomicIntegers don't even seem to be that slow, so let's just
leave this here for now.