#7 Adds the concept of a 'default language Locale'.
Soon we'll add the ability to set a language Locale using a new
constructor. For now, this provides a mechanism for declaring a
default language Locale when one is not provided. A German user might
like to set the default locale to Locale.GERMAN, for example, and can
do so via a system property.